# url-shortener/docs/swagger.yaml
definitions:
  models.CreateURLRequest:
    properties:
      url:
        description: Añadimos validación básica
        type: string
    required:
    - url
    type: object
  models.URL:
    properties:
      accessCount:
        description: Usamos int64 para contadores potencialmente grandes
        type: integer
      createdAt:
        type: string
      id:
        description: _id es el estándar en MongoDB
        type: string
      shortCode:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  models.UpdateURLRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
info:
  contact: {}
paths:
  /{shortCode}:
    get:
      description: Redirige al usuario a la URL original usando el código corto
      parameters:
      - description: Código corto
        in: path
        name: shortCode
        required: true
        type: string
      responses:
        "301":
          description: Moved Permanently
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: URL corta no encontrada
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      summary: Redirige a la URL original
      tags:
      - URLs
  /shorten/:
    post:
      consumes:
      - application/json
      description: Genera un código corto para una URL original y la almacena en la
        base de datos
      parameters:
      - description: URL a acortar
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.URL'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crea una URL corta
      tags:
      - URLs
  /shorten/{shortCode}:
    delete:
      description: Elimina la URL corta asociada a un código
      parameters:
      - description: Código corto
        in: path
        name: shortCode
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Elimina una URL corta
      tags:
      - URLs
    get:
      description: Devuelve los datos de la URL original asociada a un código corto
        (no redirige)
      parameters:
      - description: Código corto
        in: path
        name: shortCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.URL'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtiene datos de la URL original
      tags:
      - URLs
    put:
      consumes:
      - application/json
      description: Actualiza la URL original asociada a un código corto
      parameters:
      - description: Código corto
        in: path
        name: shortCode
        required: true
        type: string
      - description: Nueva URL
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.URL'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualiza la URL original
      tags:
      - URLs
  /shorten/{shortCode}/stats:
    get:
      description: Devuelve el número de accesos y datos de la URL corta
      parameters:
      - description: Código corto
        in: path
        name: shortCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.URL'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtiene estadísticas de una URL corta
      tags:
      - URLs
swagger: "2.0"